arrange the code & add register handle function that sends post request to backend
