+++ backend/routes/routes.js
@@ -10,6 +10,10 @@ router.get("/", (req, res) => {
     res.sendFile(views_dir + "index.html");
 });
 
+router.post("/new_user", (req, res) => {
+    console.log(req.body);
+})
+
 router.get("/test", (req, res) => {
     db_funcs.getAllUsers().then(query => {
         res.send(query);
