+++ frontend/controller/index.js
@@ -1,70 +1,35 @@
 function refreshPage() {
     window.location.reload();
 }
-document.getElementById('loginBtn').addEventListener('click', function() {
-  window.location.href = './../view/login.html';
-});
- document.getElementById('homebtn').addEventListener('click', function() {
-    window.location.href = './../view/index.html';
- });
- document.getElementById('registerbtn').addEventListener('click', function() {
-    window.location.href = './../view/Register.html';
-});
- document.getElementById('console').addEventListener('click', function() {
-    window.location.href = './../view/console.html';
-});
-  document.getElementById('allGames').addEventListener('click', function() {
-    window.location.href = './../view/games.html';
-  });
+
   
-  function comparePasswords(event) {
-    event.preventDefault(); // Prevent form submission
 
-    const password = document.getElementById("password").value;
-    const confirmPassword = document.getElementById("confirmPassword").value;
-    if (password === confirmPassword) {
-      // Passwords match, proceed with registration logic
-      var firstName = document.getElementById('firstName').value;
-      var lastName = document.getElementById('lastName').value;
+// $(function() {
+//     var availableTags = [
+//       "Sony playstations",
+//       "Xbox",
+//       "Nintendo",
+//       "Games",
+//       "Xbox games",
+//       "Nintendo games",
+//       "Sony playstations games",
+//       "computer games",
+//       "Console",
+//     ];
+//     $("#search").autocomplete({
+//       source: availableTags
+//     });
+//   });
 
-      // Add your registration code here using the firstName and lastName variables
+//   function handleSearch(event) {
+//     event.preventDefault(); // Prevent form submission
 
-      alert('Registration successful!');
-      window.location.href = 'success.html'
-    }
-    if (password !== confirmPassword) {
-      alert("Passwords do not match!");
-    }
-  
-    // Passwords match, proceed with registration
-    // Add your registration logic here
-    console.log("Registration successful");
-} 
+//     // Get the search query from the input field
+//     console.log("Search query:", searchQuery);
 
-$(function() {
-    var availableTags = [
-      "Sony playstations",
-      "Xbox",
-      "Nintendo",
-      "Games",
-      "Xbox games",
-      "Nintendo games",
-      "Sony playstations games",
-      "computer games",
-      "Console",
-    ];
-    $("#search").autocomplete({
-      source: availableTags
-    });
-  });
-  function handleSearch(event) {
-    event.preventDefault(); // Prevent form submission
+//     window.location.href = "search-results.html?q=" + encodeURIComponent(searchQuery);
+// }
 
-    // Get the search query from the input field
-    console.log("Search query:", searchQuery);
-
-    window.location.href = "search-results.html?q=" + encodeURIComponent(searchQuery);
-}
-var searchnameInput = document.getElementById("search");
-var searchnameValue = searchnameInput.value;
-console.log(searchnameValue);
+// var searchnameInput = document.getElementById("search");
+// var searchnameValue = searchnameInput.value;
+// console.log(searchnameValue);
