+++ frontend/model/model.js
@@ -1,0 +1,1 @@
+let addresses = []
