+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/Wahalez/dddm#readme",
   "dependencies": {
+    "body-parser": "^1.20.2",
     "dotenv": "^16.1.4",
     "express": "^4.18.2",
     "mongodb": "^5.6.0",
